forwardauth: add provided X-Forwarded headers
Add headers that the middleware provides to the target forward-auth address.
These headers contain certain request properties from the initial request that
is subject for authentication.

Without this information it is not clear, what a forward-auth target application 
may expect to receive. Having this in the docs helps application developers 
for targets of this middleware.